chore: move test-utils to own file and move functions to own file
